requirements.txt
streamlit==1.26.0
pandas==1.5.3
numpy==1.24.4
Pillow==9.4.0
scikit-learn==1.3.1
joblib==1.3.2
requests==2.31.0
rdkit-pypi==2022.9.5